Make qa-monitor tolerant to a failed build

+++ qa-monitor.py
@@ -91,12 +91,13 @@ class Repository:
 
         for n in range(MAX_DEPTH):
             commit_hash = self.log[n]['commit_hash']
-            if commit_hash in results:
+            output_path = '../%s/%s.xml' % (data_directory, commit_hash)
+            if commit_hash in results and os.path.exists(
+                    output_path.replace('..', '.')):
                 tester.load_data(commit_hash, results[commit_hash])
             else:
                 print "Testing commit " + commit_hash
                 _exec_command(self.directory, 'git', 'checkout', commit_hash)
-                output_path = '../%s/%s.xml' % (data_directory, commit_hash)
                 # le quito un punto del principio porque el path es relativo a
                 # ./scripts/
                 if not os.path.exists(output_path.replace('..', '.')):
@@ -110,11 +111,14 @@ class Repository:
         json.dump(results, open(collectors_data_path, 'w'))
 
         # convert the last findbugs file to text
-        out = _exec_command(
-            '.', '../tools/findbugs-3.0.1/bin/convertXmlToText',
-            './%s/%s.xml' % (data_directory, self.log[0]['commit_hash']))
-        with open('%s/last_report.txt' % data_directory, 'w') as f:
-            f.write(out)
+        xml_data_path = './%s/%s.xml' % (data_directory,
+                                         self.log[0]['commit_hash'])
+        if os.path.exists(xml_data_path):
+            out = _exec_command(
+                '.', '../tools/findbugs-3.0.1/bin/convertXmlToText',
+                xml_data_path)
+            with open('%s/last_report.txt' % data_directory, 'w') as f:
+                f.write(out)
 
 
 class Tester:
@@ -133,9 +137,12 @@ class Tester:
 
     def execute_processors(self, directory, output_path):
         for processor in self.processors:
-            out = _exec_command(directory, '../scripts/' + processor,
-                                output_path, directory)
-            print out
+            try:
+                out = _exec_command(directory, '../scripts/' + processor,
+                                    output_path, directory)
+                print out
+            except:
+                print "Error running processor"
 
     def execute_collectors(self, directory, commit_hash, input_path):
         self.commits.append(commit_hash)
@@ -357,11 +364,13 @@ class HtmlBuilder:
         findbugs_file = './data_%s/%s.xml' % (self.project_name,
                                               self.tester.commits[0])
         destination = '%s/%s.xml' % (output_path, self.project_name)
-        shutil.copyfile(findbugs_file, destination)
+        if os.path.exists(findbugs_file):
+            shutil.copyfile(findbugs_file, destination)
 
         txt_report_file = './data_%s/last_report.txt' % (self.project_name)
         destination = '%s/%s.txt' % (output_path, self.project_name)
-        shutil.copyfile(txt_report_file, destination)
+        if os.path.exists(txt_report_file):
+            shutil.copyfile(txt_report_file, destination)
 
     def point_of_interest_report(self, collector, commits, poi_array):
         # commits: the list of commits in the graph,
